fix: change priority attribute type from Int64 to Int32 in smart groups and mobile device groups

+++ internal/services/smart_computer_group/data_source_read.go
@@ -88,7 +88,7 @@ func (d *smartComputerGroupFrameworkDataSource) Read(ctx context.Context, req da
 	for _, criterion := range resource.Criteria {
 		criteriaModel := smartComputerGroupCriteriaDataModel{
 			Name:       types.StringValue(criterion.Name),
-			Priority:   types.Int64Value(int64(criterion.Priority)),
+			Priority:   types.Int32Value(int32(criterion.Priority)),
 			AndOr:      types.StringValue(strings.ToLower(criterion.AndOr)),
 			SearchType: types.StringValue(criterion.SearchType),
 			Value:      types.StringValue(criterion.Value),
